Add inlinetypes build flag
This change introduces a configure arg `--enable-inlinetypes`
that builds the JDK with value types enabled.

The current working (public) name for the valuetypes feature
is inlinetypes. For this reason the option should
use this name. The name for this feature has changed
in the past and may change again in the future so
the internal name "valuetypes" should remain as it is.

Depends on https://github.com/eclipse/openj9/pull/8934

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>